# backend/tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "BlockInput": {
        "properties": {
          "html": {
            "title": "Html",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "html"
        ],
        "title": "BlockInput",
        "type": "object"
      },
      "BlockResponse": {
        "properties": {
          "blockType": {
            "title": "Blocktype",
            "type": "string"
          },
          "html": {
            "title": "Html",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "orderNo": {
            "title": "Orderno",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderNo",
          "blockType",
          "html"
        ],
        "title": "BlockResponse",
        "type": "object"
      },
...
      "BookContentResponse": {
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "title": "Chapters",
            "type": "array"
          },
          "contentRating": {
            "default": "ALL",
            "title": "Contentrating",
            "type": "string"
          },
          "contentRatingDetail": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentratingdetail"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isPreview": {
            "default": false,
            "title": "Ispreview",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "priceAmt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priceamt"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "kind",
          "language",
          "status",
          "chapters"
        ],
        "title": "BookContentResponse",
        "type": "object"
      },
...
      "ChapterInput": {
        "properties": {
          "blocks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BlockInput"
            },
            "title": "Blocks",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ChapterInput",
        "type": "object"
      },
...
      "ChapterResponse": {
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockResponse"
            },
            "title": "Blocks",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "orderNo": {
            "title": "Orderno",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "title",
          "orderNo",
          "blocks"
        ],
        "title": "ChapterResponse",
        "type": "object"
      },
...
      "CreateBookRequest": {
        "properties": {
          "kind": {
            "default": "BOOK",
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "default": "ko",
            "title": "Language",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateBookRequest",
        "type": "object"
      },
      "CreateBookResponse": {
        "properties": {
          "bookId": {
            "format": "uuid",
            "title": "Bookid",
            "type": "string"
          }
        },
        "required": [
          "bookId"
        ],
        "title": "CreateBookResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportTextRequest": {
        "properties": {
          "chapterTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chaptertitle"
          },
          "rawText": {
            "title": "Rawtext",
            "type": "string"
          }
        },
        "required": [
          "rawText"
        ],
        "title": "ImportTextRequest",
        "type": "object"
      },
      "ImportTextResponse": {
        "properties": {
          "blockCount": {
            "title": "Blockcount",
            "type": "integer"
          },
          "chapterId": {
            "format": "uuid",
            "title": "Chapterid",
            "type": "string"
          }
        },
        "required": [
          "chapterId",
          "blockCount"
        ],
        "title": "ImportTextResponse",
        "type": "object"
      },
...
      "SetContentRequest": {
        "description": "\uc5d0\ub514\ud130 \uc6d0\ud074\ub9ad \ucd9c\ud310 \u2014 \uc815\ubcf8 \uc804\uccb4 \uad50\uccb4.",
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterInput"
            },
            "title": "Chapters",
            "type": "array"
          }
        },
        "required": [
          "chapters"
        ],
        "title": "SetContentRequest",
        "type": "object"
      },
...
      "SetPreviewLimitRequest": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          }
        },
        "required": [
          "limit"
        ],
        "title": "SetPreviewLimitRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "\ud55c\uc904 ebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books": {
      "post": {
        "operationId": "create_book_api_books_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Book",
        "tags": [
          "books"
        ]
      }
    },
...
    "/api/books/{book_id}/content": {
      "get": {
        "operationId": "get_content_api_books__book_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content",
        "tags": [
          "books"
        ]
      },
      "put": {
        "description": "\uc5d0\ub514\ud130 \uc6d0\ud074\ub9ad \ucd9c\ud310 \u2014 \uc815\ubcf8 \uc804\uccb4\ub97c \ud5e4\ub529 \uae30\uc900 \ucc55\ud130\ub85c \uad50\uccb4. \uc791\uac00 \ubcf8\uc778\ub9cc.",
        "operationId": "set_content_api_books__book_id__content_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Content",
        "tags": [
          "books"
        ]
      }
    },
...
    "/api/books/{book_id}/epub": {
      "get": {
        "description": "\uc815\ubcf8 \u2192 EPUB 3 \ud30c\uc77c \ub2e4\uc6b4\ub85c\ub4dc (\uc11c\uc810 \uc720\ud1b5\u00b7\uc18c\uc7a5\uc758 \uae30\ubcf8 \uc0b0\ucd9c\ubb3c).\n\n\u26a0\ufe0f \ud68c\uadc0\uac00\ub4dc: \uc774\uc804\uc5d4 \uc778\uc99d\u00b7\uad6c\ub9e4 \ud655\uc778\uc774 \uc804\ud600 \uc5c6\uc5b4 book_id(\uc2a4\ud1a0\uc5b4 URL\uc5d0 \ub178\ucd9c)\ub9cc \uc54c\uba74 \ub204\uad6c\ub098\n\ubb34\ub8cc\ub85c \uc804\uccb4 EPUB\uc744 \ubc1b\uc744 \uc218 \uc788\uc5c8\ub2e4(2026-07-08 \uc5f0\ub839\uac8c\uc774\ud2b8 \uac10\uc0ac \uc911 \ubc1c\uacac). get_content\uc758\nis_free/owned \ud310\uc815\uc744 /content \uc5d4\ub4dc\ud3ec\uc778\ud2b8\uc640 \ub3d9\uc77c\ud558\uac8c \uc801\uc6a9 \u2014 \ubbf8\ub9ac\ubcf4\uae30 \uac1c\ub150\uc774 \uc5c6\ub294 \uacbd\ub85c\ub77c\n\ubbf8\uad6c\ub9e4\u00b7\uc720\ub8cc\uba74 \uc804\uba74 \ucc28\ub2e8(NotPurchased). BookNotFound\u00b7AgeVerificationRequired\u00b7NotPurchased\n\u2192 \uc911\uc559 \ud578\ub4e4\ub7ec.",
        "operationId": "download_epub_api_books__book_id__epub_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Epub",
        "tags": [
          "books"
        ]
      }
    },
    "/api/books/{book_id}/import": {
      "post": {
        "description": "\uc6d0\uace0 import \u2014 \uc18c\uc720\uc790 \uc788\ub294 \ucc45\uc740 \uc791\uac00 \ubcf8\uc778\ub9cc(\ub0a8\uc758 \ucc45\uc5d0 \uc7a5 \ucd94\uac00 \ucc28\ub2e8).\n\nBookNotFound 404\u00b7NotOwner 403 \u2192 \uc911\uc559 \ud578\ub4e4\ub7ec\uac00 \ub9e4\ud551.",
        "operationId": "import_text_api_books__book_id__import_post",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportTextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Text",
        "tags": [
          "books"
        ]
      }
    },
...
    "/api/books/{book_id}/preview-limit": {
      "put": {
        "description": "\ubb34\ub8cc \ubbf8\ub9ac\ubcf4\uae30 \uacf5\uac1c \ubd84\ub7c9(\ube14\ub85d \uc218) \uc124\uc815 \u2014 \uc791\uac00 \ubcf8\uc778\ub9cc. 404/403 \u2192 \uc911\uc559 \ud578\ub4e4\ub7ec.",
        "operationId": "set_preview_limit_api_books__book_id__preview_limit_put",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPreviewLimitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Preview Limit",
        "tags": [
          "books"
        ]
      }
    },
...
    "/api/books/{book_id}/suggest-blurb": {
      "get": {
        "description": "\ubcf8\ubb38 \uae30\ubc18 \uc18c\uac1c\ubb38 \ucd94\ucc9c (\uc791\uac00\uac00 \uac80\ud1a0 \ud6c4 \uc800\uc7a5). BookNotFound \u2192 404 (\uc911\uc559 \ud578\ub4e4\ub7ec).",
        "operationId": "suggest_blurb_endpoint_api_books__book_id__suggest_blurb_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Suggest Blurb Endpoint Api Books  Book Id  Suggest Blurb Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Blurb Endpoint",
        "tags": [
          "books"
        ]
      }
    },
...
  }
}